feat(TT-19): renamed files, moved hook to provider and added new type
