--- src/lib/auth/context.tsx
@@ -1,14 +1,0 @@
-import React, { useState } from "react";
-import { AuthContextType, AuthProviderProps, User } from "./types";
-
-export const AuthContext = React.createContext<AuthContextType>(
-  {} as AuthContextType,
-);
-
-export const AuthProvider = ({ children }: AuthProviderProps) => {
-  const [user, setUser] = useState<User | null>(null);
-
-  const value = { user, setUser };
-
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-};
