+++ src/lib/auth/index.ts
@@ -1,2 +1,1 @@
-export * from "./hooks";
-export * from "./context";
+export * from "./provider";
