+++ src/lib/auth/provider.tsx
@@ -5,15 +5,21 @@ import {
   signInWithPopup,
   signOut,
 } from "firebase/auth";
-import { useContext, useEffect } from "react";
-import { AuthContext } from "@/lib/auth/context";
+import React, { useContext, useEffect, useState } from "react";
 import app from "@/lib/firebase";
+import { AuthContextType, AuthProviderProps, User } from "./types";
 
 const auth = getAuth(app);
 const googleAuthProvider = new GoogleAuthProvider();
 
-export const useAuth = () => {
-  const { user, setUser } = useContext(AuthContext);
+export const AuthContext = React.createContext<AuthContextType>(
+  {} as AuthContextType,
+);
+
+export const useAuth = () => useContext(AuthContext);
+
+export const AuthProvider = ({ children }: AuthProviderProps) => {
+  const [user, setUser] = useState<User | null>(null);
 
   useEffect(() => {
     return onAuthStateChanged(auth, (user) => {
@@ -32,9 +38,7 @@ export const useAuth = () => {
 
   const logOut = () => signOut(auth);
 
-  return {
-    logInGoogle,
-    logOut,
-    user,
-  };
+  const value = { user, setUser, logInGoogle, logOut };
+
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 };
